+++ sqlalchemy_utils/__init__.py
@@ -74,7 +74,7 @@ from .types import (
 from .models import Timestamp
 
 
-__version__ = '0.26.7'
+__version__ = '0.26.8'
 
 
 __all__ = (